Safari-609.1.11.1/LayoutTests/js/dom/math-random-initial-values-expected.txt

This tests early Math.random values are not consistently nearly identical.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS framesWithSimiliarEarlyRandomValues <= 5 is true
PASS successfullyParsed is true

TEST COMPLETE
